layers: Fix for VkBufferImageCopy::bufferOffset validation
According to the Vulkan spec:

"If the the calling command’s VkImage parameter’s format is not a depth/stencil
format, then VkBufferImageCopy::bufferOffset must be a multiple of the format’s
element size"

Currently the image validation layer is checking unconditionally if
VkBufferImageCopy::bufferOffset is a multiple of the format’s element size
and reporting error otherwise.

Added code that skips the alignment check if the VkImage parameter’s format is
combined depth/stencil.